Scaffold organization (#1)
* scaffold the dev, prod workspaces

* remove the 'cs-' prefix from all google cloud generated mmodules

* clean up over-complexity in global dir, use resource naming convention in projects

see section 5 (resource hierarhy) for some brief exaplanation on the shared project set up and suggested naming convention

* remove the logging project and related resources temporarily

* change to state backend to bucket in project that resides within the dpgraham.com organization

* copy existing modules from dpgraham-server repo

* setup initial dev terraform

* add project name to variables in order to differentiate project ID and name

* add vpc module for dev environment

* move modules from modules sub-directory which was not supposed to be there

* rename module to serverless-load-balancer, scaffold new global-load-balancer module

the serverless module creates a load balanacer that is not flexible enough to use with services hosted using VM, autoscalers, k8, etc. and only works with serverless services. for now, we're just going to create a new module

* add gcp-apis module for enabling APIs, POC create cloud sql instance

* update sql module and initial usage in dev environment

* use variables to set set up the development database

* update global terraform module documentation on setting up

* use tfvars for varialbes in global scope

* remove creating vpc in dev project

* add vpc creation back to project

will see about connecting google managed services such as SQL to a project local VPC then if need be we can use VPC peering to connect to connect to the shared VPC

* add github boilerplate workflows, code of conduct, PR template, contribution guidelines

* add artifact regsitry to dev environment

* remove artifact registry default input values

* (not working) artificat registry path is messed up

* remove cloudsql.admin from gcp-devops group at organizational level and add to development folder level

* clean up the iam policy terraform configs

we now use a naming convention to easily group the IAM polices by principal and resource, which has the added benefit of condensing the file a bit and making ti easier to read. Before, every policy had a separate terraform module instance block

* remove internal IP for now

* initial subnet to separate resource creation, setup as shared vpc subnet in development environment

* move subnet creation from inline to separate resources for dev

* add shared prod vpc to prod environment project

* add subnets to prod and add to list of shared subnets

* quick fix to rename resources with underscores instead of hyphens

* using shared VPC subnets in environments

so following an example article found online at https://medium.com/google-cloud/how-to-create-shared-vpc-with-terraform-on-google-cloud-c66e4130abea, we were able to successfully follow along. I've learned that we can't connect a cloud SQL instance to the shared VPC but I can setup VMs on the shared VPC

* update sql module, remove old databases

* simplify the sql module by remove the VPC connector, use the shared VPC instead